Hack in support for abstract sockets
I'm reverse engineering a program which uses abstract sockets. These
are sockets which don't correspond to the filesystem, they instead
live in a single global namespace (WITH NO SECURITY) and the trick is
that the `sun_path` _starts_ with a `'\0'` character.

Most tools (e.g. netstat) will reference abstracts sockets by starting
with an `@` character.

This change adds some more code duplication. Can be done better!